MAVEN-138 create a simpler base mojo class that doesn't automatically download portal-web dependencies
