+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/LiferayMojo.java
@@ -1,0 +1,50 @@
+/*******************************************************************************
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ *
+ *******************************************************************************/
+package com.liferay.maven.plugins;
+
+import org.apache.maven.plugin.AbstractMojo;
+import org.apache.maven.project.MavenProject;
+
+
+/**
+ * @author Gregory Amerson
+ */
+public abstract class LiferayMojo extends AbstractMojo
+{
+	protected boolean isLiferayProject() {
+		String packaging = project.getPackaging();
+
+		if (packaging.equals("pom")) {
+			getLog().info("Skipping " + project.getArtifactId());
+
+			return false;
+		}
+
+		return true;
+	}
+
+	/**
+	 * @parameter expression="${liferayVersion}"
+	 */
+	protected String liferayVersion;
+
+	/**
+	 * @parameter expression="${project}"
+	 * @required
+	 * @readonly
+	 */
+	protected MavenProject project;
+
+}
